Fix imports for server subpackage
