+++ agent/server/__init__.py
@@ -1,0 +1,5 @@
+"""WebSocket server package."""
+
+from .websocket import AgentWebSocketServer
+
+__all__ = ["AgentWebSocketServer"]
